# sandbox-api/docs/openapi.yml
openapi: 3.0.0
info:
  contact: {}
  description: API for manipulating filesystem, processes and network.
  title: Sandbox API
  version: 0.0.1
paths:
...
  /filesystem-multipart:
    get:
      description: List all active multipart uploads
      responses:
        "200":
          description: List of active uploads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipartListUploadsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: List multipart uploads
      tags:
        - filesystem
  "/filesystem-multipart/{uploadId}/abort":
    delete:
      description: Abort a multipart upload and clean up all parts
      parameters:
        - description: Upload ID
          in: path
          name: uploadId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload aborted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Abort multipart upload
      tags:
        - filesystem
  "/filesystem-multipart/{uploadId}/complete":
    post:
      description: Complete a multipart upload by assembling all parts
      parameters:
        - description: Upload ID
          in: path
          name: uploadId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultipartCompleteRequest"
        description: List of parts
        required: true
      responses:
        "200":
          description: Upload completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Complete multipart upload
      tags:
        - filesystem
  "/filesystem-multipart/{uploadId}/part":
    put:
      description: Upload a single part of a multipart upload
      parameters:
        - description: Upload ID
          in: path
          name: uploadId
          required: true
          schema:
            type: string
        - description: Part number (1-10000)
          in: query
          name: partNumber
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: Part data
                  type: string
                  format: binary
              required:
                - file
        required: true
      responses:
        "200":
          description: Part uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipartUploadPartResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Upload part
      tags:
        - filesystem
  "/filesystem-multipart/{uploadId}/parts":
    get:
      description: List all uploaded parts for a multipart upload
      parameters:
        - description: Upload ID
          in: path
          name: uploadId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of parts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipartListPartsResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: List parts
      tags:
        - filesystem
...
  /lsp:
    get:
      description: Get a list of all active LSP servers
      responses:
        "200":
          description: List of LSP servers
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LSPServerResponse"
                type: array
      summary: List all LSP servers
      tags:
        - lsp
    post:
      description: "Create a new LSP server for a specific language and project. Supported languages: python, typescript, javascript"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLSPServerRequest"
        description: LSP server creation request
        required: true
      responses:
        "200":
          description: LSP server information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LSPServerResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Create an LSP server
      tags:
        - lsp
  "/lsp/{id}":
    delete:
      description: Shutdown and remove an LSP server
      parameters:
        - description: LSP server ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: LSP server deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: LSP server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Delete an LSP server
      tags:
        - lsp
    get:
      description: Get information about a specific LSP server
      parameters:
        - description: LSP server ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: LSP server information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LSPServerResponse"
        "404":
          description: LSP server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Get LSP server information
      tags:
        - lsp
  "/lsp/{id}/completions":
    post:
      description: Get code completion suggestions from an LSP server
      parameters:
        - description: LSP server ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionRequest"
        description: Completion request
        required: true
      responses:
        "200":
          description: Completion items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: LSP server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Get code completions
      tags:
        - lsp
  "/network/process/{pid}/monitor":
    delete:
      description: Stop monitoring for new ports opened by a process
      parameters:
        - description: Process ID
          in: path
          name: pid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Object containing PID and success message
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid process ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Stop monitoring ports for a process
      tags:
        - network
    post:
      description: Start monitoring for new ports opened by a process
      parameters:
        - description: Process ID
          in: path
          name: pid
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortMonitorRequest"
        description: Port monitor configuration
        required: true
      responses:
        "200":
          description: Object containing PID and success message
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Start monitoring ports for a process
      tags:
        - network
  "/network/process/{pid}/ports":
    get:
      description: Get a list of all open ports for a process
      parameters:
        - description: Process ID
          in: path
          name: pid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Object containing PID and array of network.PortInfo
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid process ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Get open ports for a process
      tags:
        - network
  /network/tunnel:
    delete:
      description: "Stop the network tunnel and restore the original network configuration. WARNING: After disconnecting, the sandbox will lose all outbound internet connectivity (no egress). Inbound connections to the sandbox will still work. Use PUT /network/tunnel/config to re-establish the tunnel."
      responses:
        "200":
          description: Tunnel disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: No tunnel is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to stop tunnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Disconnect tunnel
      tags:
        - network
  /network/tunnel/config:
    put:
      description: Apply a new tunnel configuration on the fly. The existing tunnel is torn down and a new one is established. This endpoint is write-only; there is no corresponding GET to read the config back.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TunnelConfigRequest"
        description: Base64-encoded tunnel configuration
        required: true
      responses:
        "200":
          description: Configuration applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid tunnel configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to apply configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Update tunnel configuration
      tags:
        - network
  /process:
    get:
      description: Get a list of all running and completed processes
      responses:
        "200":
          description: Process list
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ProcessResponse"
                type: array
      summary: List all processes
      tags:
        - process
    post:
      description: Execute a command and return process information. If Accept header is text/event-stream, streams logs in SSE format and returns the process response as a final event.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRequest"
        description: Process execution request
        required: true
      responses:
        "200":
          description: Process information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ProcessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Execute a command
      tags:
        - process
  "/process/{identifier}":
    delete:
      description: Gracefully stop a running process
      parameters:
        - description: Process identifier (PID or name)
          in: path
          name: identifier
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Process not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Stop a process
      tags:
        - process
    get:
      description: Get information about a process by its PID or name
      parameters:
        - description: Process identifier (PID or name)
          in: path
          name: identifier
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResponse"
        "404":
          description: Process not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Get process by identifier
      tags:
        - process
  "/process/{identifier}/kill":
    delete:
      description: Forcefully kill a running process
      parameters:
        - description: Process identifier (PID or name)
          in: path
          name: identifier
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process killed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Process not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Kill a process
      tags:
        - process
...
servers:
  - url: https://sbx-{sandbox_id}-{workspace_id}.{region}.bl.run
    variables:
      sandbox_id:
        default: unknown
      workspace_id:
        default: unknown
      region:
        default: unknown
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      properties:
        error:
          example: Error message
          type: string
      required:
        - error
      type: object
...
    LSPServerResponse:
      properties:
        createdAt:
          example: Wed, 01 Jan 2023 12:00:00 GMT
          type: string
        errorMsg:
          example: ""
          type: string
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        languageId:
          example: python
          type: string
        processPid:
          example: 12345
          type: integer
        projectPath:
          example: /workspace/project
          type: string
        status:
          example: ready
          type: string
      required:
        - createdAt
        - id
        - languageId
        - processPid
        - projectPath
        - status
      type: object
    MultipartCompleteRequest:
      properties:
        parts:
          items:
            $ref: "#/components/schemas/MultipartPartInfo"
          type: array
      type: object
...
    MultipartListPartsResponse:
      properties:
        parts:
          items:
            $ref: "#/components/schemas/filesystem.UploadedPart"
          type: array
        uploadId:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
      type: object
    MultipartListUploadsResponse:
      properties:
        uploads:
          items:
            $ref: "#/components/schemas/filesystem.MultipartUpload"
          type: array
      type: object
...
    MultipartUploadPartResponse:
      properties:
        etag:
          example: 5d41402abc4b2a76b9719d911017c592
          type: string
        partNumber:
          example: 1
          type: integer
        size:
          example: 5242880
          type: integer
      type: object
...
    SuccessResponse:
      properties:
        message:
          example: File created successfully
          type: string
        path:
          example: /path/to/file
          type: string
      required:
        - message
      type: object
...
security:
  - BearerAuth: []